Politicas de acceso para proteger la API
Se añade la verificacion 'assertJsonApiError' a los test 'guests_cannot_delete_appointments' y 'guests_cannot_update_appointments'.

Se añade el test 'cannot_update_appointments_owned_by_other_users' para verificar que no se puedan modificar appointments que no le pertenezcan al usuario (que le pertenezcan a otros usuarios).

Una de las opciones para realizar la autorizacion es dentro del Request (SaveAppointmentRequest en este caso) en donde se verifica si el id del usuario que se encuentre logueado en el momento (request()->user()->id) es igual al id del autor del appoinment que se quiere modificar.

public function authorize(): bool
{
    return ->user()->id === ->route('appointment')->author->id;
}

La validacion tambien se puede hacer dentro de cada metodo del controlador que se necesite.

public function update(SaveAppointmentRequest , Appointment ): AppointmentResource
{
    if (->user()->id !== ->author->id) {
        throw new AccessDeniedHttpException;
    }

    ->update(->validated());

    return AppointmentResource::make();
}

Sin embargo, de la manera que se quiere hacer, es con politicas de acceso.

php artisan make:policy AppointmentPolicy -m Appointment

Estas, se ubican dentro de App\Policies. Laravel automaticamente detecta que si se tiene una politica llamada AppointmentPolicy se va a adjuntar directamente al modelo Appointment, es decir, el nombre del modelo, seguido de la palabra 'Policy', si se sigue esta convencion, no se debe modificar nada adicional para que funcione correctamente.

Dentro del metodo en el controlador se puede hacer uso de ->authorize('update', ) recibiendo como parametro el nombre de la politica y luego el modelo, en este caso, se esta verificando si el usuario autenticado puede actualizar el appointment, se ejecuta el metodo en la politica y es aqui donde se determina si puede hacerlo o no, detras de escenas, en caso de que no se pueda actualizar, Laravel automaticamente va a mandar un error 403 (No autorizado).

Se usa la funcion is de laravel, que recibe como paramtro un modelo con el cual se quiere comparar si es igual con aquel donde se llamó la funcion.

+++ app/Http/Controllers/Api/AppointmentController.php
@@ -4,12 +4,14 @@ namespace App\Http\Controllers\Api;
 
 use App\Models\Appointment;
 
+use Illuminate\Http\Request;
 use Illuminate\Http\Response;
 use App\Http\Controllers\Controller;
 use App\Http\Resources\AppointmentResource;
 use App\Http\Requests\SaveAppointmentRequest;
 use Illuminate\Http\Resources\Json\JsonResource;
 use Illuminate\Http\Resources\Json\AnonymousResourceCollection;
+use Symfony\Component\HttpKernel\Exception\AccessDeniedHttpException;
 
 class AppointmentController extends Controller
 {
@@ -63,6 +65,11 @@ class AppointmentController extends Controller
      */
     public function update(SaveAppointmentRequest $request, Appointment $appointment): AppointmentResource
     {
+        /** Se manda como primer parametro el nombre del metodo de la politica
+         * Policy (que se encuentra en la ruta App\Policies), y como segundo el modelo.
+         */
+        $this->authorize('update', $appointment);
+
         $appointment->update($request->validated());
 
         return AppointmentResource::make($appointment);
@@ -71,8 +78,13 @@ class AppointmentController extends Controller
     /**
      * Remove the specified resource from storage.
      */
-    public function destroy(Appointment $appointment): Response
+    public function destroy(Appointment $appointment, Request $request): Response
     {
+        /** Se manda como primer parametro el nombre del metodo de la politica
+         * Policy (que se encuentra en la ruta App\Policies), y como segundo el modelo.
+         */
+        $this->authorize('delete', $appointment);
+
         $appointment->delete();
 
         return response()->noContent();
